# Conflicts:
#	webapp/hss_H5/src/main/java/com/jkm/hss/controller/wx/LoginController.java
#	webapp/hss_H5/src/main/webapp/WEB-INF/jsp/material.jsp

+++ service/dealer/dealer-api/src/main/java/com/jkm/hss/dealer/service/DealerService.java
@@ -489,7 +489,7 @@ public interface DealerService {
      * @param firstLevelDealerId
      * @return
      */
-    String selectProxyName(int firstLevelDealerId);
+    String selectProxyName(long firstLevelDealerId);
 
     /**
      * 根据一级代理商id查询名称
